Hateoas, Jackson, Binder - 4.1 ~ 5.1

+++ spring-web-mvc-intermediario-rest-jax/src/main/java/br/com/robierto/springwebmvcintermediariorestjax/repository/SoldadoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import br.com.robierto.springwebmvcintermediariorestjax.entity.SoldadoEntity;
 
-public interface SoldadoRepository extends CrudRepository<SoldadoEntity, Integer>{
+public interface SoldadoRepository extends CrudRepository<SoldadoEntity, Long>{
 	List<SoldadoEntity> findAll();
 }
